+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 0.9.10
 
+- Runes are now craftable out of stone tool materials
 - Add italian translation thanks to Zano1999 #3
+- Add `runes:runes` tag, including all runes
+- Add Rune Crafting Alter block item tooltip hint
 
 # 0.9.9